Seperated all websocket logic into layers, entity, routes, and services. Start file is now renamed to 'main.py'
