+++ backend/logic/api/config.py
@@ -1,0 +1,1 @@
+BOARD_COUNT = 2
